Fixed some review comments
